+++ crates/eframe/src/epi.rs
@@ -223,6 +223,13 @@ pub struct NativeOptions {
     /// Default: `false`.
     pub fullscreen: bool,
 
+    /// On Mac: the window doesn't have a titlebar, but floating window buttons.
+    ///
+    /// See [winit's documentation][with_fullsize_content_view] for information on Mac-specific options.
+    ///
+    /// [with_fullsize_content_view]: https://docs.rs/winit/latest/x86_64-apple-darwin/winit/platform/macos/trait.WindowBuilderExtMacOS.html#tymethod.with_fullsize_content_view
+    pub fullsize_content: bool,
+
     /// On Windows: enable drag and drop support. Drag and drop can
     /// not be disabled on other platforms.
     ///
@@ -361,6 +368,7 @@ impl Default for NativeOptions {
             maximized: false,
             decorated: true,
             fullscreen: false,
+            fullsize_content: false,
             drag_and_drop_support: true,
             icon_data: None,
             initial_window_pos: None,